refactor: Migrate users.create endpoint to OpenAPI with AJV validation
