+++ .changeset/migrate-users-create-openapi.md
@@ -1,0 +1,6 @@
+---
+"@rocket.chat/meteor": minor
+"@rocket.chat/rest-typings": minor
+---
+
+Add OpenAPI support for the `users.create` API endpoint by migrating to chained route definition with AJV body and response validation.
