+++ packages/rest-typings/src/v1/users.ts
@@ -3,7 +3,6 @@ import type { IExportOperation, ISubscription, ITeam, IUser, IPersonalAccessToke
 import { ajv } from './Ajv';
 import type { PaginatedRequest } from '../helpers/PaginatedRequest';
 import type { PaginatedResult } from '../helpers/PaginatedResult';
-import type { UserCreateParamsPOST } from './users/UserCreateParamsPOST';
 import type { UserDeactivateIdleParamsPOST } from './users/UserDeactivateIdleParamsPOST';
 import type { UserLogoutParamsPOST } from './users/UserLogoutParamsPOST';
 import type { UserRegisterParamsPOST } from './users/UserRegisterParamsPOST';
@@ -271,12 +270,6 @@ export type UsersEndpoints = {
 		};
 	};
 
-	'/v1/users.create': {
-		POST: (params: UserCreateParamsPOST) => {
-			user: IUser;
-		};
-	};
-
 	'/v1/users.update': {
 		POST: (params: UsersUpdateParamsPOST) => {
 			user: IUser;
@@ -370,7 +363,6 @@ export type UsersEndpoints = {
 	};
 };
 
-export * from './users/UserCreateParamsPOST';
 export * from './users/UserSetActiveStatusParamsPOST';
 export * from './users/UserDeactivateIdleParamsPOST';
 export * from './users/UsersInfoParamsGet';
