--- packages/rest-typings/src/v1/users/UserCreateParamsPOST.ts
@@ -1,52 +1,0 @@
-import type { IUserSettings } from '@rocket.chat/core-typings';
-
-import { ajv } from '../Ajv';
-
-export type UserCreateParamsPOST = {
-	email: string;
-	name: string;
-	password: string;
-	username: string;
-	active?: boolean;
-	bio?: string;
-	nickname?: string;
-	statusText?: string;
-	roles?: string[];
-	joinDefaultChannels?: boolean;
-	requirePasswordChange?: boolean;
-	setRandomPassword?: boolean;
-	sendWelcomeEmail?: boolean;
-	verified?: boolean;
-	customFields?: Record<string, any>;
-	settings?: IUserSettings;
-	freeSwitchExtension?: string;
-	/* @deprecated */
-	fields: string;
-};
-
-const userCreateParamsPostSchema = {
-	type: 'object',
-	properties: {
-		email: { type: 'string' },
-		name: { type: 'string' },
-		password: { type: 'string' },
-		username: { type: 'string' },
-		active: { type: 'boolean', nullable: true },
-		bio: { type: 'string', nullable: true },
-		nickname: { type: 'string', nullable: true },
-		statusText: { type: 'string', nullable: true },
-		roles: { type: 'array', items: { type: 'string' } },
-		joinDefaultChannels: { type: 'boolean', nullable: true },
-		requirePasswordChange: { type: 'boolean', nullable: true },
-		setRandomPassword: { type: 'boolean', nullable: true },
-		sendWelcomeEmail: { type: 'boolean', nullable: true },
-		verified: { type: 'boolean', nullable: true },
-		customFields: { type: 'object' },
-		fields: { type: 'string', nullable: true },
-		freeSwitchExtension: { type: 'string', nullable: true },
-	},
-	additionalProperties: false,
-	required: ['email', 'name', 'password', 'username'],
-};
-
-export const isUserCreateParamsPOST = ajv.compile<UserCreateParamsPOST>(userCreateParamsPostSchema);
